Added react-router and created basic rotues
