+++ client/src/Css/App.css
@@ -88,34 +88,32 @@ body {
   font-family: "poppins", "roboto", sans-serif;
 }
 
-#root {
-  height: 100%;
-}
-
 img {
   max-width: 100%;
   display: block;
 }
 
-.gallery {
+#root {
   height: 100%;
   display: grid;
-  grid-template-columns: repeat(12, 1fr);
-  grid-template-rows: 10% 10% auto 20%;
+  grid-template-columns: 1fr;
+  grid-template-rows: 10% auto 20%;
+}
+
+.content {
+  grid-column: 1/2;
+  grid-row: 2/3;
+  display: grid;
+  place-items: center;
 }
 
 .header {
-  grid-column: 1/13;
+  grid-column: 1/2;
   grid-row: 1/2;
 }
 
-.filter_container {
-  grid-column: 4/10;
-  grid-row: 2/3;
-}
-
 .photo_grid {
-  grid-column: 1/13;
+  grid-column: 1/2;
   grid-row: 3/4;
   justify-self: center;
   margin: 1em;
@@ -131,8 +129,8 @@ img {
 }
 
 .footer {
-  grid-column: 1/13;
-  grid-row: 4/5;
+  grid-column: 1/2;
+  grid-row: 3/4;
 }
 
 .lightbox-container {
