+++ client/src/Sass/base/_reset.scss
@@ -27,11 +27,6 @@ body {
   }
 }
 
-//div that contains react app
-#root {
-  height: 100%;
-}
-
 //reset for images
 img {
   max-width: 100%;
