--- client/src/Sass/base/gridLayout.scss
@@ -1,37 +1,0 @@
-.gallery {
-  height: 100%;
-  display: grid;
-  grid-template-columns: repeat(12, 1fr);
-  grid-template-rows: 10% 10% auto 20%;
-}
-
-.header {
-  grid-column: 1/13;
-  grid-row: 1/2;
-}
-
-.filter_container {
-  grid-column: 4/10;
-  grid-row: 2/3;
-}
-
-.photo_grid {
-  grid-column: 1/13;
-  grid-row: 3/4;
-  justify-self: center;
-  margin: 1em;
-}
-
-.loading-container {
-  grid-column: 1/13;
-  grid-row: 2/4;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
-.footer {
-  grid-column: 1/13;
-  grid-row: 4/5;
-}
