+++ client/src/index.js
@@ -4,10 +4,13 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import App from "./App.jsx";
-import reportWebVitals from "./reportWebVitals";
+import { BrowserRouter } from "react-router-dom";
+
 ReactDOM.render(
   <React.StrictMode>
-    <App />
+    <BrowserRouter>
+      <App />
+    </BrowserRouter>
   </React.StrictMode>,
   document.getElementById("root")
 );
