+++ client/src/pages/Settings.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function Settings() {
+  return <div>Settings</div>;
+}
+
+export default Settings;
