+++ client/src/pages/View.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function View() {
+  return <div>View</div>;
+}
+
+export default View;
